requirements.txt
selenium
seleniumbase
python-dotenv
faker